Move way properties to OSM
