+++ src/ext/java/org/opentripplanner/ext/reportapi/model/BicycleSafetyReport.java
@@ -1,7 +1,7 @@
 package org.opentripplanner.ext.reportapi.model;
 
-import org.opentripplanner.graph_builder.module.osm.WayPropertySet;
 import org.opentripplanner.graph_builder.module.osm.tagmapping.OsmTagMapper;
+import org.opentripplanner.openstreetmap.wayproperty.WayPropertySet;
 
 public class BicycleSafetyReport {
