+++ src/main/java/org/opentripplanner/graph_builder/module/osm/tagmapping/AtlantaMapper.java
@@ -1,9 +1,9 @@
 package org.opentripplanner.graph_builder.module.osm.tagmapping;
 
-import static org.opentripplanner.graph_builder.module.osm.WayPropertiesBuilder.withModes;
+import static org.opentripplanner.openstreetmap.wayproperty.WayPropertiesBuilder.withModes;
 import static org.opentripplanner.street.model.StreetTraversalPermission.ALL;
 
-import org.opentripplanner.graph_builder.module.osm.WayPropertySet;
+import org.opentripplanner.openstreetmap.wayproperty.WayPropertySet;
 
 /**
  * OSM way properties for the Atlanta, Georgia, USA area.
