+++ src/main/java/org/opentripplanner/openstreetmap/spi/OSMProvider.java
@@ -1,8 +1,8 @@
 package org.opentripplanner.openstreetmap.spi;
 
 import java.time.ZoneId;
-import org.opentripplanner.graph_builder.module.osm.WayPropertySet;
 import org.opentripplanner.graph_builder.module.osm.tagmapping.OsmTagMapper;
+import org.opentripplanner.openstreetmap.wayproperty.WayPropertySet;
 
 public interface OSMProvider {
   ZoneId getZoneId();
