+++ src/main/java/org/opentripplanner/openstreetmap/wayproperty/LocalizedStringMapper.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.graph_builder.module.osm;
+package org.opentripplanner.openstreetmap.wayproperty;
 
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ListMultimap;
@@ -9,7 +9,6 @@ import java.util.Objects;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import org.opentripplanner.framework.resources.ResourceBundleSingleton;
-import org.opentripplanner.graph_builder.module.osm.tagmapping.DefaultMapper;
 import org.opentripplanner.openstreetmap.model.OSMWithTags;
 import org.opentripplanner.transit.model.basic.I18NString;
 import org.opentripplanner.transit.model.basic.LocalizedString;
@@ -33,15 +32,16 @@ class LocalizedStringMapper {
   /**
    * Creates String which can be localized.
    * <p>
-   * Uses {@link #getTagNames() } to get which tag values are needed for this key. For each of this
-   * tag names tag value is read from OSM way. If tag value is missing it is added as empty string.
+   * Uses {@link #getTagNames(String) } to get which tag values are needed for this key. For each of
+   * this tag names tag value is read from OSM way. If tag value is missing it is added as empty
+   * string.
    * <p>
    * For example. If key platform has key {ref} current value of tag ref in way is saved to be used
    * in localizations. It currently assumes that tag exists in way. (otherwise this namer wouldn't
    * be used)
    * </p>
    *
-   * @param key key of translation for this way set in {@link DefaultMapper} and translations read
+   * @param key key of translation for this way set in the WayPropertyMapper and translations read
    *            from properties Files
    * @param way OSM way from which tag values are read
    */
