+++ src/main/java/org/opentripplanner/openstreetmap/wayproperty/NoteProperties.java
@@ -1,8 +1,7 @@
-package org.opentripplanner.graph_builder.module.osm;
+package org.opentripplanner.openstreetmap.wayproperty;
 
 import java.util.Map;
 import java.util.regex.Pattern;
-import org.opentripplanner.graph_builder.module.osm.tagmapping.DefaultMapper;
 import org.opentripplanner.model.NoteMatcher;
 import org.opentripplanner.model.StreetNote;
 import org.opentripplanner.model.StreetNoteAndMatcher;
@@ -11,9 +10,6 @@ import org.opentripplanner.openstreetmap.model.OSMWithTags;
 import org.opentripplanner.transit.model.basic.I18NString;
 import org.opentripplanner.transit.model.basic.TranslatedString;
 
-/**
- * Currently unused since notes are disabled in {@link DefaultMapper}
- */
 public class NoteProperties {
 
   private static final Pattern patternMatcher = Pattern.compile("\\{(.*?)}");
