+++ src/main/java/org/opentripplanner/openstreetmap/wayproperty/SafetyFeatures.java
@@ -1,4 +1,4 @@
-package org.opentripplanner.graph_builder.module.osm;
+package org.opentripplanner.openstreetmap.wayproperty;
 
 /**
  * Record that holds forward and back safety factors for cycling or walking.
